+++ src/onegov.org/onegov/org/tests/test_views.py
@@ -4056,6 +4056,11 @@ def test_directory_submissions(client, postgres):
     # change it to accept submissions
     page = page.click("Konfigurieren")
     page.form['enable_submissions'] = True
+    page = page.form.submit()
+
+    # this fails because there are invisible fields
+    assert "«Description» nicht sichtbar" in page
+    page.form['lead_format'] = '[Description]'
     page = page.form.submit().follow()
 
     assert "Eintrag vorschlagen" in page
@@ -4130,11 +4135,10 @@ def test_directory_submissions(client, postgres):
     page.form['structure'] = """
         Name *= ___
         Description = ...
-        Category *=
-            [ ] Monument
-            [ ] Vista
+        Category *= ___
     """
-    page.form.submit()
+    page.form['lead_format'] = '[Description] [Category]'
+    page.form.submit().follow()
 
     client.post(accept_url)
 
@@ -4143,7 +4147,7 @@ def test_directory_submissions(client, postgres):
 
     # in which case we can edit the submission to get it up to snuff
     page = page.click("Details bearbeiten")
-    page.select_checkbox('category', "Monument")
+    page.form['category'] = 'Monument'
     page.form.submit().follow()
 
     client.post(accept_url)
@@ -4181,6 +4185,53 @@ def test_directory_submissions(client, postgres):
 
     assert '\n' not in desc
     transaction.abort()
+
+
+def test_directory_change_requests(client):
+
+    client.login_admin()
+
+    # create a directory that accepts change requests
+    page = client.get('/directories').click('Verzeichnis')
+    page.form['title'] = "Playgrounds"
+    page.form['structure'] = """
+        Name *= ___
+    """
+    page.form['enable_change_requests'] = True
+    page.form['title_format'] = '[Name]'
+    page = page.form.submit().follow()
+
+    # create an entry
+    page = page.click('Eintrag')
+    page.form['name'] = 'Central Park'
+    page = page.form.submit().follow()
+
+    # ask for a change
+    page = page.click("Änderung vorschlagen")
+    page.form['name'] = 'Diana Ross Playground'
+    page.form['submitter'] = 'user@example.org'
+    page.form['comment'] = 'This is better'
+    page.form.submit().follow().form.submit().follow()
+
+    # check the ticket
+    page = client.get('/tickets/ALL/open').click("Annehmen").follow()
+    assert '<del>Central Park</del><ins>Diana Ross Playground</ins>' in page
+    assert 'This is better' in page
+
+    # make sure it hasn't been applied yet
+    assert 'Central Park' in \
+        client.get('/directories/playgrounds/central-park')
+
+    # apply the changes
+    page.click("Übernehmen")
+    page = client.get(page.request.url)
+    assert 'Central Park' not in page
+    assert 'Diana Ross Playground' in page
+    assert 'This is better' in page
+
+    # check if they were applied (the id won't have changed)
+    assert 'Diana Ross Playground' in \
+        client.get('/directories/playgrounds/central-park')
 
 
 def test_dependent_number_form(client):
@@ -4531,13 +4582,13 @@ def test_markdown_in_directories(client):
     """
     page.form['title_format'] = '[Name]'
     page.form['content_fields'] = 'Notes'
-    page.form.submit()
+    page.form.submit().follow()
 
     page = client.get('/directories/clubs')
     page = page.click('Eintrag', index=0)
     page.form['name'] = 'Soccer Club'
     page.form['notes'] = '* Soccer rules!'
-    page.form.submit()
+    page.form.submit().follow()
 
     assert "<li>Soccer rules" in client.get('/directories/clubs/soccer-club')